% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_session.R
\name{create_session}
\alias{create_session}
\title{Create or Update Session Skeleton}
\usage{
create_session(..., package)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=session_opt]{session_opt}}
  \describe{
    \item{\code{opts}}{a list: the reference options to use. Unspecified options will
take the default values specified for other arguments}
    \item{\code{shorthand}}{a string: the session shorthand name}
    \item{\code{fullname}}{a string: the session full name}
    \item{\code{tutorials}}{a character vector: if non-zero length, creates
corresponding tutorial file skeletons}
    \item{\code{practicals}}{a character vector: if non-zero length, creates
corresponding practical script file skeletons}
    \item{\code{data}}{a character vector: if non-zero length, creates a session data
file for documentation.}
  }}

\item{package}{the package directory}
}
\description{
Create or Update Session Skeleton
}
