% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_topic.R
\name{create_topic}
\alias{create_topic}
\title{Create or Update Topic Skeleton}
\usage{
create_topic(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=topic_opt]{topic_opt}}
  \describe{
    \item{\code{opts}}{a list: the reference options to use. Unspecified options will
take the default values specified for other arguments}
    \item{\code{shorthand}}{a string: the topic shorthand name}
    \item{\code{name}}{a string: the topic full name}
    \item{\code{tutorials}}{a character vector: if non-zero length, creates
corresponding tutorial file skeletons}
    \item{\code{practicals}}{a character vector: if non-zero length, creates
corresponding practical script file skeletons}
    \item{\code{data}}{a character vector: if non-zero length, creates a topic data
file for documentation.}
  }}
}
\description{
Create or Update Topic Skeleton
}
