% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_opt.R
\name{session_opt}
\alias{session_opt}
\title{Create Session Options}
\usage{
session_opt(
  opts = list(),
  shorthand = "DUMMY",
  fullname = "NOT FOR DUMMIES",
  tutorials = c(),
  practicals = c(),
  data = c()
)
}
\arguments{
\item{opts}{a list: the reference options to use. Unspecified options will
take the default values specified for other arguments}

\item{shorthand}{a string: the session shorthand name}

\item{fullname}{a string: the session full name}

\item{tutorials}{a character vector: if non-zero length, creates
corresponding tutorial file skeletons}

\item{practicals}{a character vector: if non-zero length, creates
corresponding practical script file skeletons}

\item{data}{a character vector: if non-zero length, creates a session data
file for documentation.}
}
\value{
a list, matching the argument requirements for \code{\link[=create_clinic]{create_clinic()}} and
\code{\link[=create_session]{create_session()}}.
}
\description{
Creates a standardized options object for sessions templating.
}
