% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_opt.R
\name{topic_opt}
\alias{topic_opt}
\title{Create Topic Options}
\usage{
topic_opt(
  opts = list(),
  shorthand = "DUMMY",
  fullname = shorthand,
  tutorials = c(),
  practicals = c(),
  data = c(),
  overwrite = FALSE
)
}
\arguments{
\item{opts}{a list: the reference options to use. Unspecified options will
take the default values specified for other arguments}

\item{shorthand}{a string: the topic shorthand name}

\item{fullname}{a string: the topic full name}

\item{tutorials}{a character vector: if non-zero length, creates
corresponding tutorial file skeletons}

\item{practicals}{a character vector: if non-zero length, creates
corresponding practical script file skeletons}

\item{data}{a character vector: if non-zero length, creates a topic data
file for documentation.}
}
\value{
a list, matching the argument requirements for \code{\link[=create_clinic]{create_clinic()}} and
\code{\link[=create_topic]{create_topic()}}.
}
\description{
Creates a standardized options object for topic templating.
}
