% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_practical.R
\name{use_practical}
\alias{use_practical}
\title{Create a Practical}
\usage{
use_practical(name, topic)
}
\arguments{
\item{name}{a string: the title of the practical; after sanitization, it will
be the initial filename for the practical}

\item{topic}{a string: the shorthand topic title}
}
\description{
Creates a practical script skeleton, associated with a
particular package topic, in the current package.
}
