% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tutorial.R
\name{use_tutorial}
\alias{use_tutorial}
\title{Create a Session Tutorial}
\usage{
use_tutorial(name, sessionname)
}
\arguments{
\item{name}{a string: the title of the tutorial; after sanitization, will
also be the initial filename for the tutorial.}

\item{sessionname}{a string: the shorthand session title}
}
\description{
Creates a \code{learnr} tutorial skeleton, associated with a
particular package topic, in the current package.
}
