% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_practical.R, R/use_tutorial.R
\name{use_practical}
\alias{use_practical}
\alias{use_activity}
\alias{use_tutorial}
\title{Add Learning Activites}
\usage{
use_practical(name, topic, overwrite = FALSE, open = rlang::is_interactive())

use_tutorial(name, topic, overwrite = FALSE, open = rlang::is_interactive())
}
\arguments{
\item{name}{a string: the title of the activity; after sanitization, will
also be the initial file name.}

\item{topic}{a string: the shorthand topic title}

\item{overwrite}{logical: overwrite existing materials?}

\item{open}{logical: open the resulting file?}
}
\description{
Creates \code{learnr} tutorials or R script skeletons,
associated with a particular topic.
}
