% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_practical.R
\name{use_practical}
\alias{use_practical}
\title{Create a Session Practical}
\usage{
use_practical(name, sessionname)
}
\arguments{
\item{name}{a string: the title of the tutorial; after sanitization, it will
be the initial filename for the tutorial.}

\item{sessionname}{a string: the shorthand session title}
}
\description{
Creates a practical script skeleton, associated with a
particular package topic, in the current package.
}
