% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tutorial.R
\name{use_tutorial}
\alias{use_tutorial}
\title{Add a Tutorial}
\usage{
use_tutorial(name, topic)
}
\arguments{
\item{name}{a string: the title of the tutorial; after sanitization, will
also be the initial filename for the tutorial.}

\item{topic}{a string: the shorthand topic title}
}
\description{
Creates a \code{learnr} tutorial skeleton in the current package,
associated with a particular topic.
}
